refactor: own unit for graphics
